% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cause_params_sims.R
\name{cause_params_sims}
\alias{cause_params_sims}
\title{Estimate CAUSE parameters for simulated data}
\usage{
cause_params_sims(dat, null_wt = 10, no_ld = FALSE,
  max_candidates = Inf)
}
\arguments{
\item{dat}{A simulated data frame created with sum_stats}

\item{null_wt}{Null weight in dirichlet prior on mixing parameters}

\item{no_ld}{Run with the nold data (T/F)}
}
\description{
Estimate CAUSE parameters for simulated data
}
