% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cause_sims.R
\name{get_sigma}
\alias{get_sigma}
\title{Helper function for simulations testing prior on gamma and eta}
\usage{
get_sigma(effect, q)
}
\description{
Given an effect and a quantile, q, find the value of sigma_g
such that pnorm(effect, 0, sigma_g) = q
}
