% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsmr_sims.R
\name{gsmr_sims}
\alias{gsmr_sims}
\title{Run GSMR on simulated data}
\usage{
gsmr_sims(dat, evd_list, p_val_thresh = 5e-08, no_ld = FALSE)
}
\arguments{
\item{dat}{Data frame of simulated data}

\item{evd_list}{List of eigen value decompositions for LD regions}

\item{no_ld}{Use no LD data}

\item{p_value_thresh}{p-value threshold}
}
\description{
Run GSMR on simulated data
}
