% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcv_sims.R
\name{lcv_sims}
\alias{lcv_sims}
\title{Run LCV with simulated data}
\usage{
lcv_sims(dat, no_ld = FALSE, intercepts = !no_ld, sig.thresh = 30)
}
\arguments{
\item{dat}{Simulated data set}

\item{no_ld}{Use no LD data}

\item{sig.thresh}{Signicance threshold for LCV to estimate crosstrait.intercept and ldsc.intercept}

\item{intercpets}{use crosstrait.interecpt=1 and ldsc.intercept=1 in RunLCV}
}
\description{
Run LCV with simulated data
}
