% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal_mr.R
\name{MBE}
\alias{MBE}
\title{Modal MR}
\usage{
MBE(BetaXG, BetaYG, seBetaXG, seBetaYG, phi = c(1, 0.5, 0.25),
  n_boot = 10000, alpha = 0.05)
}
\arguments{
\item{BetaXG}{vector of instrument-exposure regression coefficients}

\item{BetaYG}{vector of instrument-outcome regression coefficients}

\item{seBetaXG}{vector of instrument-exposure standard errors (SEs)}

\item{seBetaYG}{vector of instrument-outcome SEs}

\item{phi}{tunning parameter (e.g., 1=default bandwidth; 0.5=half of the default bandwidth)}

\item{n_boot}{number of bootstrap iterations}

\item{alpha}{alpha level of the confidence intervals (e.g., alpha=0.05 corresponds to 1-0.05=95\% confidence intervals)}
}
\description{
Implementation of the Modal MR method from Hartwig et al 2017.
This code is copied from the supplement of that paper.
}
