% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpresso.R
\name{mrpresso}
\alias{mrpresso}
\title{Run MR-PRESSO on simulated data}
\usage{
mrpresso(dat, p_val_thresh = 5e-08, no_ld = FALSE)
}
\arguments{
\item{dat}{A simulated data frame created with sum_stats}

\item{p_val_thresh}{p-value threshold}

\item{no_ld}{Run with the nold data (T/F)}
}
\description{
Run MR-PRESSO on simulated data
}
