% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_data.R
\name{reverse_data}
\alias{reverse_data}
\title{Reverse a simulated data set}
\usage{
reverse_data(dat, evd_list, ld_prune_pval_thresh = 0.001, r2_thresh = 0.1)
}
\arguments{
\item{dat}{Data set simulated with sum_stats}

\item{evd_list}{List of eigen value decompositions for LD regions}

\item{ld_prune_pval_thresh}{Minimum p-value to be included in LD pruning}

\item{r2_thres}{r^2 threshold for LD pruning}
}
\description{
Reverse a simulated data set
}
