% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunLCV.R
\name{RunLCV}
\alias{RunLCV}
\title{LCV}
\usage{
RunLCV(ell, z.1, z.2, no.blocks = 100, crosstrait.intercept = 1,
  ldsc.intercept = 1, weights = 1/pmax(1, ell),
  sig.threshold = .Machine$integer.max, n.1 = 1, n.2 = 1,
  intercept.12 = 0)
}
\arguments{
\item{ell}{Mx1 vector of LD scores;}

\item{z.1}{Mx1 vector of estimated marginal per-normalized-genotype effects on trait 1    (or Z scores; invariant to scaling);}

\item{z.2}{Mx2 vector of effects on trait2;}

\item{no.blocks, }{number of jackknife blocks;}

\item{crosstrait.intercept}{0 if crosstrait LDSC intercept should be fixed and 1 otherwise;}

\item{ldsc.intercept}{0 if LDSC intercept should be fixed and 1 otherwise (1 is recommended);}

\item{weights}{Mx1 vector of regression weights;}

\item{sig.threshold, }{threshold above which to discard chisq statistics for the purpose of estimating
the LDSC intercept; large-effect SNPs discarded above sig_threshold*mean(z.x^2);}

\item{intercept.12}{covariance between sampling errors for Z1 and Z2, only needed if
crosstrait_intercept=0 If the 2 GWAS are disjoint, this can be set to zero.}

\item{n1}{sample size for trait 1, only needed if ldsc.intercept=1;}

\item{n2}{sample size for trait 2, only needed if ldsc.intercept=1;}
}
\description{
Main function for running LCV. Obtained from https://github.com/lukejoconnor/LCV
}
