% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_sims.R
\name{setup_sims}
\alias{setup_sims}
\title{Set up simulation directory}
\usage{
setup_sims()
}
\description{
A helper function that prepares a working directory for running simulations.
}
\details{
This function prepares a working directory to replicate the simulations presented in Morrison et al 2019
(https://www.biorxiv.org/content/10.1101/682237v3). After running this function you will be able to execute simulation
experiments using DSC using the commands

 nohup dsc --replicate 100 --host config.yml -c 4 [dscfile] > [outputfile] &


 You will need to modify the config.yml file to work with your compute cluster.
}
