% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_stats.R
\name{sum_stats}
\alias{sum_stats}
\title{Simulate summary statistics}
\usage{
sum_stats(snps, evd_list, n_copies = 30, n1, n2, h1, h2, neffect1,
  neffect2, gamma, eta, q, tau, omega, cores,
  ld_prune_pval_thresh = 0.001, r2_thresh = 0.1)
}
\arguments{
\item{evd_list}{List of eigen value decompositions for LD regions}

\item{n_copies}{How many times to replicate chromosome 19}

\item{n1}{Sample size in study 1}

\item{n2}{sample size in study 2}

\item{h1}{Expected heritability of trait 1}

\item{h2}{Expected heritability of trait 2}

\item{neffect1}{Average number of trait 1 effect snps}

\item{neffect2}{Average number of trait 2 effect snps. Note that this is in addition to SNPs that effect Y through M.}

\item{gamma}{gamma}

\item{eta}{eta}

\item{q}{q}

\item{tau}{tau = h1*(gamma^2)/h2 Proportion of Y heritability explained by M through causal mechanism}

\item{omega}{omega = h1*(eta)^2/h2 omega*q^2 is proportion of Y heritability explained by M through confounder mechanism}

\item{snsp}{SNP info data frame}
}
\description{
Simulate summary statistics
}
