% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_stats_nold.R
\name{sum_stats_nold}
\alias{sum_stats_nold}
\title{Simulate summary statistics, nold}
\usage{
sum_stats_nold(p, n1, n2, h1, h2, neffect1, neffect2, gamma, eta, q, tau,
  omega)
}
\arguments{
\item{p}{Total number of snps}

\item{n1}{Sample size in study 1}

\item{n2}{sample size in study 2}

\item{h1}{Expected heritability of trait 1}

\item{h2}{Expected heritability of trait 2}

\item{neffect1}{Average number of trait 1 effect snps}

\item{neffect2}{Average number of trait 2 effect snps. Note that this is in addition to SNPs that effect Y through M.}

\item{gamma}{gamma}

\item{eta}{eta}

\item{q}{q}

\item{tau}{tau = h1*(gamma^2)/h2 Proportion of Y heritability explained by M through causal mechanism}

\item{omega}{omega = h1*(eta)^2/h2 omega*q^2 is proportion of Y heritability explained by M through confounder mechanism}
}
\description{
Simulate summary statistics, nold
}
