% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cause_sims.R
\name{cause_sims}
\alias{cause_sims}
\title{Run CAUSE on simulated data}
\usage{
cause_sims(dat, params, sigma_g, qalpha = 1, qbeta = 10,
  no_ld = FALSE, thresh = 0.001)
}
\arguments{
\item{dat}{A simulated data frame created with sum_stats}

\item{params}{A cause params object. You can make this with cause_params_sims}

\item{sigma_g}{Prior variance for gamma and eta. If left missing will be estimated.}

\item{qalpha, qbeta}{parameters defining prior distribution on q. q~Beta(qalpha,qbeta)}

\item{no_ld}{Run with the nold data (T/F)}

\item{thresh}{p-value threshold}
}
\description{
Run CAUSE on simulated data
}
